requirements.txt
pillow
url-normalize
matplotlib
requests
tqdm
scipy